Fix: keep args dump task metadata header lightweight
Host-build-graph runtime types only need the dump selection POD, but including the full args_dump header pulled platform timing symbols into orchestration sources and conflicted with paged-attention definitions.

Move the task metadata and its ABI assertions into a dependency-light header, then include that header from both a2a3 and a5 runtime types. Paged-attention scene compilation and dump-args level 1/3 simulation tests pass on the supported platforms.
